added beans
