+++ src/main/java/ru/otus/spring/hw01/services/IQuestionReader.java
@@ -1,4 +1,4 @@
-package services;
+package ru.otus.spring.hw01.services;
 
 import org.apache.commons.csv.CSVRecord;
