+++ src/main/java/ru/otus/spring/hw01/services/IUserName.java
@@ -1,4 +1,4 @@
-package services;
+package ru.otus.spring.hw01.services;
 
 public interface IUserName {
     String getUserName();
