+++ src/main/java/ru/otus/spring/hw01/services/UserNameImpl.java
@@ -1,4 +1,4 @@
-package services;
+package ru.otus.spring.hw01.services;
 
 import java.util.Scanner;
